Внесла корректировки урок № 8 в задании нужно массив сделать из 5 фигур,(сделала: круг, треугольник, квадрат, прямоугольник, трапеция.
