--- src/lesson08/Figure.java
@@ -1,6 +1,0 @@
-package lesson08;
-
-public interface Figure {
-    void squareFigure();
-    int perimeterFigure();
-}
