+++ src/lesson08/Part1/Figure.java
@@ -1,0 +1,6 @@
+package lesson08.Part1;
+
+public interface Figure { // класс интерфейс, который содержит 2 метода: для подсчета периметра и площади
+    void squareFigure();
+    int perimeterFigure();
+}
