+++ src/lesson08/Part2/Employed.java
@@ -1,4 +1,4 @@
-package lesson08;
+package lesson08.Part2;
 
 public interface Employed {
     void printPosition (String position);
